requirements.txt
# Core
python-telegram-bot==21.0
python-dotenv==1.0.0
pyyaml==6.0.1

# Database
sqlalchemy==2.0.27
alembic==1.13.1

# LLM Integration
httpx==0.27.0
openai==1.12.0  # For OpenRouter compatibility

# Scheduling
apscheduler==3.10.4

# Utilities
pydantic==2.6.1
jsonschema==4.21.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0

# Logging
loguru==0.7.2
